requirements.txt
onnxruntime
numpy
pillow
tqdm